Moved compiler binary to a separate crate

+++ src/common/types.rs
@@ -3,14 +3,14 @@ use std::fmt::{Display, Formatter};
 #[derive(Debug)]
 pub enum Type {
     Bool,
-    Int64,
+    Int32,
 }
 
 impl Display for Type {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         match self {
             Type::Bool => write!(f, "bool"),
-            Type::Int64 => write!(f, "i64"),
+            Type::Int32 => write!(f, "i32"),
         }
     }
 }
